Trading_App/requirements.txt
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.2
decorator==4.4.1
defusedxml==0.6.0
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JSGlue==0.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
MarkupSafe==1.1.1
parso==0.8.0
Pillow==7.1.2
pycparser==2.20
six==1.14.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
WTForms==2.3.1
